# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'OpenForum - API documentation'
  description: 'OpenForum default API documentation'
  version: 1.0.0
servers:
  - url: 'http(s)://[current-host]'
paths:
...
  /api/users/email/verify/request-new:
    post:
      summary: 'Request new verification email'
      operationId: requestNewVerificationEmail
      description: 'Send a new email verification link once requested or return the current user instance if already verified'
      parameters:
        - in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        - in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        - in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: 'Verification email sent'
          content:
            application/json:
              schema:
                type: object
                example:
                  data: [ ]
                  meta:
                    success: true
                    errors: null
                properties:
                  data:
                    type: array
                    example: [ ]
                    description: 'Container for resource.'
                  meta:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                        description: 'Response status summary, if `true` the call was successful.'
                      errors:
                        type: string
                        example: null
                        description: 'null|object[] If one or more errors occur this field will be filled with errors.'
      tags:
        - 'User management'
  '/api/users/email/verify/{id}/{hash}':
    post:
      summary: 'Verify email'
      operationId: verifyEmail
      description: "Handle the email verification, forwarding the verification call to Fortify's default controller."
      parameters:
        - in: query
          name: expires
          description: 'expiration timestamp of the verification link.'
          example: 1676736522
          required: true
          schema:
            type: integer
            description: 'expiration timestamp of the verification link.'
            example: 1676736522
        - in: query
          name: signature
          description: 'Url signature hash.'
          example: 2e99a9278bcba67a7b52acbb6f2fb822c72c6d4b1d63fb059033439059d9f250
          required: true
          schema:
            type: string
            description: 'Url signature hash.'
            example: 2e99a9278bcba67a7b52acbb6f2fb822c72c6d4b1d63fb059033439059d9f250
        - in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        - in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        - in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    type: user
                    id: usr_10833034c4dd47849af8ed8ea4a68a1b
                    attributes:
                      name: 'Devonte Goldner'
                      email: 'imo********************'
                      email_verified_at: '18/02/2023 11:05:51'
                      token: 2|aFGHN9Ox9Uc42XfLqrXlY2xAzNwhQLDb7kIf3rjN
                    relationships:
                      reservations:
                        data:
                          - type: reservation
                            id: resrv_1e302893c3cd4dcd8c7810ef1a07c180
                  included:
                    - type: reservation
                      id: resrv_1e302893c3cd4dcd8c7810ef1a07c180
                      attributes:
                        user: usr_10833034c4dd47849af8ed8ea4a68a1b
                        event: evt_d7e868f3083344f6bec2548db7979d6b
                        activity: act_4c0158279c0148439fb285ebc7ad8c12
                  meta:
                    success: true
                    errors: null
                properties:
                  data:
                    type: object
                    properties:
                      type:
                        type: string
                        example: user
                        description: 'Resource return type as of JSON:API standard, always equal to `user`.'
                      id:
                        type: string
                        example: usr_10833034c4dd47849af8ed8ea4a68a1b
                        description: 'Resource unique identifier, always starts with `usr_`.'
                      attributes:
                        type: object
                        properties:
                          name:
                            type: string
                            example: 'Devonte Goldner'
                            description: 'Current user full name.'
                          email:
                            type: string
                            example: 'imo********************'
                            description: 'Current user censored email.'
                          email_verified_at:
                            type: string
                            example: '18/02/2023 11:05:51'
                            description: "Email verification datetime or null, if this field is\n               null for a given user he won't be allowed to reserve any spot."
                          token:
                            type: string
                            example: 2|aFGHN9Ox9Uc42XfLqrXlY2xAzNwhQLDb7kIf3rjN
                            description: 'Current user authentication token.'
                      relationships:
                        type: object
                        properties:
                          reservations:
                            type: object
                            properties:
                              data:
                                type: array
                                example:
                                  - type: reservation
                                    id: resrv_1e302893c3cd4dcd8c7810ef1a07c180
                                description: 'Reservations identifier data container.'
                                items:
                                  type: object
                                  properties:
                                    type:
                                      type: string
                                      example: reservation
                                    id:
                                      type: string
                                      example: resrv_1e302893c3cd4dcd8c7810ef1a07c180
                  included:
                    type: array
                    example:
                      - type: reservation
                        id: resrv_1e302893c3cd4dcd8c7810ef1a07c180
                        attributes:
                          user: usr_10833034c4dd47849af8ed8ea4a68a1b
                          event: evt_d7e868f3083344f6bec2548db7979d6b
                          activity: act_4c0158279c0148439fb285ebc7ad8c12
                    description: 'Container for included relationships.'
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          example: reservation
                        id:
                          type: string
                          example: resrv_1e302893c3cd4dcd8c7810ef1a07c180
                        attributes:
                          type: object
                          properties:
                            user:
                              type: string
                              example: usr_10833034c4dd47849af8ed8ea4a68a1b
                            event:
                              type: string
                              example: evt_d7e868f3083344f6bec2548db7979d6b
                            activity:
                              type: string
                              example: act_4c0158279c0148439fb285ebc7ad8c12
                  meta:
                    type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                        description: 'Response status summary, if `true` the call was successful.'
                      errors:
                        type: string
                        example: null
                        description: 'null|object[] If one or more errors occur this field will be filled with errors.'
      tags:
        - 'User management'
    parameters:
      - in: path
        name: id
        description: 'The identifier of the user to verify.'
        example: 1
        required: true
        schema:
          type: integer
      - in: path
        name: hash
        description: 'Verification hash.'
        example: 567159d622ffbb50b11b0efd307be358624a26ee
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  - default: [ ]